fixed folder names and responsivenessOfProfileCard
